Make sign_android_image.sh support EROFS image format as well.
Doc: go/arcvm-erofs

BUG=b:275280979
BRANCH=None
TEST=unittests
TEST=Run with an EROFS Android image
TEST=Run an Squashfs Android image and ensure output doesn't change


Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/vboot_reference/+/4554126
Tested-by: Satoshi Niwa <niwa@chromium.org>
Commit-Queue: Satoshi Niwa <niwa@chromium.org>

+++ scripts/image_signing/sign_android_image.sh
@@ -307,7 +307,7 @@ image_content_integrity_check() {
   return 0
 }
 
-list_image_files() {
+list_files_in_squashfs_image() {
   local unsquashfs=$1
   local system_img=$2
   "${unsquashfs}" -l "${system_img}" | grep ^squashfs-root
@@ -340,9 +340,18 @@ sign_android_internal() {
     "${root_fs_dir}"/opt/google/*/android/system.raw.img)"
   local android_dir=$(dirname "${android_system_image}")
   local system_img="${android_dir}/system.raw.img"
+
   # Use the versions in $PATH rather than the system ones.
-  local unsquashfs=$(which unsquashfs)
-  local mksquashfs=$(which mksquashfs)
+  # squashfs-tools
+  local unsquashfs mksquashfs
+  unsquashfs=$(which unsquashfs)
+  mksquashfs=$(which mksquashfs)
+
+  # erofs-utils
+  local dump_erofs fsck_erofs mkfs_erofs
+  dump_erofs=$(which dump.erofs)
+  fsck_erofs=$(which fsck.erofs)
+  mkfs_erofs=$(which mkfs.erofs)
 
   if [[ $# -ne 2 ]]; then
     flags_help
@@ -353,19 +362,41 @@ sign_android_internal() {
     die "System image does not exist: ${system_img}"
   fi
 
-  # NOTE: Keep compression_flags aligned with
-  # src/private-overlays/project-cheets-private/scripts/board_specific_setup.sh
-  local compression_flags=""
-  local compression=$(sudo "${unsquashfs}" -s "${system_img}" \
-    | grep -e ^"Compression\s")
-  if [[ "${compression}" == "Compression gzip" ]]; then
-    compression_flags="-comp gzip"
-  elif [[ "${compression}" == "Compression lz4" ]]; then
-    compression_flags="-comp lz4 -Xhc -b 256K"
-  elif [[ "${compression}" == "Compression zstd" ]]; then
-    compression_flags="-comp zstd -b 256K"
+  local image_type
+  local output_squashfs output_erofs
+  if output_squashfs=$("${unsquashfs}" -s "${system_img}" 2>&1); then
+    image_type="squashfs"
+  elif output_erofs=$("${dump_erofs}" "${system_img}" 2>&1); then
+    image_type="erofs"
   else
-    die "Unexpected compression type: ${compression}"
+    die "Can't detect the image type of ARC system image\n" \
+        "output_squashfs: ${output_squashfs}\n" \
+        "output_erofs: ${output_erofs}"
+  fi
+  info "Detected ARC system image type: ${image_type}"
+
+  local squashfs_comp_flags=()
+  if [[ "${image_type}" == "squashfs" ]]; then
+    # NOTE: Keep squashfs_comp_flags aligned with
+    # project-cheets-private/scripts/board_specific_setup.sh
+    local compression
+    compression="$("${unsquashfs}" -s "${system_img}" \
+      | awk '$1 == "Compression" {print $2}')"
+    case "${compression}" in
+      "gzip")
+        squashfs_comp_flags=(-comp gzip)
+        ;;
+      "lz4")
+        squashfs_comp_flags=(-comp lz4 -Xhc -b 256K)
+        ;;
+      "zstd")
+        squashfs_comp_flags=(-comp zstd -b 256K)
+        ;;
+      *)
+        die "Unexpected compression type: ${compression}"
+        ;;
+    esac
+    info "Detected squashfs compression type: ${compression}"
   fi
 
   if ! type -P zipalign &>/dev/null || ! type -P signapk &>/dev/null \
@@ -380,16 +411,28 @@ sign_android_internal() {
   local system_mnt="${working_dir}/mnt"
   local system_capabilities_orig="${working_dir}/capabilities.orig"
 
-  # Extract with xattrs so we can read and audit capabilities. See b/179170462.
-  info "Unpacking squashfs system image with xattrs to ${system_mnt}"
-  sudo "${unsquashfs}" -x -f -no-progress -d "${system_mnt}" "${system_img}"
-  snapshot_capabilities "${system_mnt}" > "${system_capabilities_orig}"
-  sudo rm -rf "${system_mnt}"
+  if [[ "${image_type}" == "squashfs" ]]; then
+    # Extract with xattrs so we can read and audit capabilities.
+    # See b/179170462.
+    info "Unpacking squashfs system image with xattrs to ${system_mnt}"
+    sudo "${unsquashfs}" -x -f -no-progress -d "${system_mnt}" "${system_img}"
+    snapshot_capabilities "${system_mnt}" > "${system_capabilities_orig}"
+    sudo rm -rf "${system_mnt}"
 
-  info "Unpacking squashfs system image without xattrs to ${system_mnt}"
-  list_image_files "${unsquashfs}" "${system_img}" > \
-      "${working_dir}/image_file_list.orig"
-  sudo "${unsquashfs}" -no-xattrs -f -no-progress -d "${system_mnt}" "${system_img}"
+    info "Unpacking squashfs system image without xattrs to ${system_mnt}"
+    list_files_in_squashfs_image "${unsquashfs}" "${system_img}" > \
+        "${working_dir}/image_file_list.orig"
+    sudo "${unsquashfs}" -no-xattrs -f -no-progress -d \
+        "${system_mnt}" "${system_img}"
+
+  elif [[ "${image_type}" == "erofs" ]]; then
+    # TODO(b/286015959): Preserve capabilities for EROFS as well.
+    info "Unpacking erofs system image to ${system_mnt}"
+    sudo "${fsck_erofs}" "--extract=${system_mnt}" "${system_img}"
+
+    # List all files inside the image.
+    find "${system_mnt}" > "${working_dir}/image_file_list.orig"
+  fi
 
   snapshot_file_properties "${system_mnt}" > "${working_dir}/properties.orig"
 
@@ -425,13 +468,19 @@ sign_android_internal() {
       local packages_before=$(grep "<package " "${packages_cache}" | wc -l)
       local vendor_mnt=$(make_temp_dir)
       local vendor_img="${android_dir}/vendor.raw.img"
-      local jar_lib="lib/arc-cache-builder/org.chromium.arc.cachebuilder.jar"
-      info "Unpacking squashfs vendor image to ${vendor_mnt}/vendor"
-      # Vendor image is not updated during this step. However we have to include
-      # vendor apks to re-generated packages cache which exists in one file for
-      # both system and vendor images.
-      sudo "${unsquashfs}" -x -f -no-progress -d "${vendor_mnt}/vendor" \
-          "${vendor_img}"
+
+      if [[ "${image_type}" == "squashfs" ]]; then
+        info "Unpacking squashfs vendor image to ${vendor_mnt}/vendor"
+        # Vendor image is not updated during this step. However we have to
+        # include vendor apks to re-generated packages cache which exists in
+        # one file for both system and vendor images.
+        sudo "${unsquashfs}" -x -f -no-progress -d "${vendor_mnt}/vendor" \
+            "${vendor_img}"
+      elif [[ "${image_type}" == "erofs" ]]; then
+        info "Unpacking erofs vendor image to ${vendor_mnt}/vendor"
+        sudo "${fsck_erofs}" "--extract=${vendor_mnt}/vendor" "${vendor_img}"
+      fi
+
       if ! arc_generate_packages_cache "${system_mnt}" "${vendor_mnt}" \
           "${working_dir}/packages_cache.xml" \
           "${working_dir}/file_hash_cache"; then
@@ -456,26 +505,45 @@ sign_android_internal() {
     info "Packages cache ${packages_cache} does not exist. Skip regeneration."
   fi
 
-  # Apply original capabilities to system image and verify correctness.
-  if ! apply_capabilities "${system_mnt}" "${system_capabilities_orig}"; then
-    return 1
-  fi
-  if ! capabilities_integrity_check "${system_mnt}" "${working_dir}"; then
-    return 1
+  if [[ "${image_type}" == "squashfs" ]]; then
+    # Apply original capabilities to system image and verify correctness.
+    # TODO(b/286015959): Support this for EROFS as well.
+    if ! apply_capabilities "${system_mnt}" "${system_capabilities_orig}"; then
+      return 1
+    fi
+    if ! capabilities_integrity_check "${system_mnt}" "${working_dir}"; then
+      return 1
+    fi
   fi
 
-  info "Repacking squashfs image with compression flags '${compression_flags}'"
   local old_size=$(stat -c '%s' "${system_img}")
   # Remove old system image to prevent mksquashfs tries to merge both images.
   sudo rm -rf "${system_img}"
-  sudo mksquashfs "${system_mnt}" "${system_img}" \
-    ${compression_flags} -context-file "${file_contexts}" -mount-point "/" \
-    -no-progress
-  local new_size=$(stat -c '%s' "${system_img}")
+
+  if [[ "${image_type}" == "squashfs" ]]; then
+    info "Repacking squashfs image with ${squashfs_comp_flags[*]}"
+    sudo "${mksquashfs}" "${system_mnt}" "${system_img}" \
+      -context-file "${file_contexts}" -mount-point "/" \
+      -no-progress "${squashfs_comp_flags[@]}"
+
+    list_files_in_squashfs_image "${unsquashfs}" "${system_img}" > \
+        "${working_dir}/image_file_list.new"
+
+  elif [[ "${image_type}" == "erofs" ]]; then
+    # List all files inside the image.
+    find "${system_mnt}" > "${working_dir}/image_file_list.new"
+
+    info "Repacking erofs image"
+    # TODO(b/286165395): Detect the original EROFS compression type and use it
+    #                    instead of always using lz4hc.
+    sudo "${mkfs_erofs}" -z lz4hc --file-contexts "${file_contexts}" \
+      "${system_img}" "${system_mnt}"
+  fi
+
+  local new_size
+  new_size=$(stat -c '%s' "${system_img}")
   info "Android system image size change: ${old_size} -> ${new_size}"
 
-  list_image_files "${unsquashfs}" "${system_img}" > \
-      "${working_dir}/image_file_list.new"
   if d=$(grep -v -F -x -f "${working_dir}"/image_file_list.{new,orig}); then
     # If we have a line in image_file_list.orig which does not appear in
     # image_file_list.new, it means some files are removed during signing
